refactor: SharedPageApi axios 적용(을 하려다가 뭔가 이상한)
