+++ src/Components/SignUp/SignUpForm.js
@@ -1,5 +1,5 @@
 import { useEffect, useRef, useState } from "react";
-import styles from "@/styles/Signin.module.css";
+import styles from "@/styles/S<ignin.module.css>";
 import Image from "next/image";
 import { useRouter } from "next/router";
 import axios from "@/lib/axios";
