+++ src/lib/axios.js
@@ -1,7 +1,9 @@
 import baseAxios from "axios";
 
+const baseURL = process.env.NEXT_PUBLIC_LINKBRARY_BASEURL;
+
 const axios = baseAxios.create({
-  baseURL: "https://bootcamp-api.codeit.kr/api/",
+  baseURL: baseURL,
 });
 
 export default axios;
